refactor: extract normalizeList into shared util and import in both pages
Agent-Logs-Url: https://github.com/UPRI-earthquake/earthquake-hub-frontend/sessions/c3dab3b9-eea8-4833-8aa5-f335bac4c225
